requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
soundfile==0.12.1
tensorflow==2.15.0
pillow==10.1.0
boto3==1.34.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
